moves tags_to_groups to separate module
