+++ packages/pytest-simcore/src/pytest_simcore/helpers/postgres_tags.py
@@ -5,7 +5,8 @@
 
 
 from aiopg.sa.connection import SAConnection
-from simcore_postgres_database.models.tags import tags, tags_to_groups
+from simcore_postgres_database.models.tags import tags
+from simcore_postgres_database.models.tags_to_groups import tags_to_groups
 
 
 async def create_tag_access(
